+++ lib/kitchen/directions/eoc_section_title_link_snippet.rb
@@ -10,7 +10,7 @@ module Kitchen
             <h3 data-type="document-title" id="#{page.title.copied_id}">
               <span class="os-number">#{chapter.count_in(:book)}.#{page.count_in(:chapter)}</span>
               <span class="os-divider"> </span>
-              <span class="os-text" data-type="" itemprop="">#{page.title.text}</span>
+              <span class="os-text" data-type="" itemprop="">#{page.title_text}</span>
             </h3>
           </a>
         HTML