+++ lib/kitchen/page_element.rb
@@ -28,6 +28,14 @@ module Kitchen
       first!(is_introduction? ? selectors.title_in_introduction_page : selectors.title_in_page)
     end
 
+    # Returns the title's text regardless of whether the title has been baked
+    #
+    # @return [String]
+    #
+    def title_text
+      title.children.one? ? title.text : title.first('.os-text').text
+    end
+
     # Returns an enumerator for titles.
     #
     # @return [ElementEnumerator]