+++ spec/directions/bake_toc_spec.rb
@@ -21,7 +21,12 @@ RSpec.describe Kitchen::Directions::BakeToc do
           </h1>
         </div>
         <div data-type="unit">
-          <h1 data-type="document-title">Unit 1 Title</h1>
+          <h1 data-type="document-title">
+            <span class="os-part-text">Unit </span>
+            <span class="os-number">1</span>
+            <span class="os-divider"> </span>
+            <span data-type="" itemprop="" class="os-text">Unit 1 Title</span>
+          </h1>
           <div data-type="chapter">
             <h1 data-type="document-title" id="chapTitle1">
               <span class="os-part-text">Chapter </span>
@@ -282,9 +287,9 @@ RSpec.describe Kitchen::Directions::BakeToc do
             </li>
             <li class="os-toc-unit" cnx-archive-shortid="" cnx-archive-uri="">
               <a href="#">
-                <span class="os-number"><span class="os-part-text">Unit </span> 1</span>
+                <span class="os-number"><span class="os-part-text">Unit </span>1</span>
                 <span class="os-divider"> </span>
-                <span class="os-text" data-type="" itemprop=""> Unit 1 Title </span>
+                <span class="os-text" data-type="" itemprop="">Unit 1 Title</span>
               </a>
               <ol class="os-unit">
                 <li class="os-toc-chapter" cnx-archive-shortid="" cnx-archive-uri="">
@@ -333,9 +338,9 @@ RSpec.describe Kitchen::Directions::BakeToc do
             </li>
             <li class="os-toc-unit" cnx-archive-shortid="" cnx-archive-uri="">
               <a href="#">
-                <span class="os-number"><span class="os-part-text">Unit </span> 2</span>
+                <span class="os-number"><span class="os-part-text">Unit </span>2</span>
                 <span class="os-divider"> </span>
-                <span class="os-text" data-type="" itemprop=""> Unit 2 Title </span>
+                <span class="os-text" data-type="" itemprop="">Unit 2 Title</span>
               </a>
               <ol class="os-unit">
                 <li class="os-toc-chapter" cnx-archive-shortid="" cnx-archive-uri="">